[GONG-50] refactor: 집중모드 페이지에서 시간 관련 핸들러 변경으로인한 레이아웃 및 컴포넌트 구조 대응 [cid] layout.tsx에서 시간 관련 핸들러 추가 시간 관련 핸들러 변경으로 인한 PeriodName, PeriodTimeline 코드 대응
